refactor(core): split domain errors to per-module files + add config views (Task #9)
Part B3: Domain error classes moved from core/errors.py to their owning
modules (identity/errors.py, sandbox/errors.py, metering/errors.py,
storage/errors.py, platform/errors.py).  core/errors.py retains the base
class and generic errors, plus backward-compat re-exports via star imports.

Part B4: Per-module config.py files added for sandbox, identity, storage,
infra, and proxy — each exposes a read-only settings view with only the
fields that module needs.  Existing code continues to work; new code can
import from the local config view.

All 806 tests pass, lint clean, OpenAPI unchanged.
